+++ src/blog.go
@@ -7,24 +7,18 @@
 package main
 
 import (
-	"bytes"
 	"flag"
 	"fmt"
-	"math"
 	"net/http"
 	"os"
-	"path"
-	"regexp"
 
+	"github.com/mariomac/goblog/src/assets"
 	"github.com/mariomac/goblog/src/install"
 	"github.com/mariomac/goblog/src/legacy"
 	"github.com/mariomac/goblog/src/logr"
 	"github.com/sirupsen/logrus"
 
-	"github.com/mariomac/goblog/src/blog"
 	"github.com/mariomac/goblog/src/conn"
-	"github.com/mariomac/goblog/src/feed"
-	"github.com/mariomac/goblog/src/visual"
 )
 
 var log *logrus.Entry
@@ -33,31 +27,6 @@ func init() {
 	// TODO: make log level configurable
 	logrus.SetLevel(logrus.DebugLevel)
 	log = logr.Get()
-}
-
-
-// Directory names
-
-
-
-
-var entries blog.Entries
-var templates = visual.Templater{}
-
-var validPagePath = regexp.MustCompile(`^([_\-a-zA-Z0-9]+)\.md$`)
-
-func makePageHandler(
-	rootPath string,
-	fn func(http.ResponseWriter, *http.Request, string),
-) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		m := validPagePath.FindStringSubmatch(r.URL.Path[len(rootPath):])
-		if m == nil {
-			http.NotFound(w, r)
-			return
-		}
-		fn(w, r, m[0])
-	}
 }
 
 func main() {
@@ -80,24 +49,15 @@ func main() {
 	log.Printf("Configuration: %#v", cfg)
 
 	log.Print("Starting GoBlog...")
-
-	// Load blog entries
-
-
-	// Create Atom XML feed
-	atomxml := bytes.NewBufferString(
-		feed.BuildAtomFeed(entries.Sorted(0, math.MaxInt), cfg.Domain, pathEntry)).Bytes()
-
-
-	mux := http.NewServeMux()
-	mux.HandleFunc(pathAtom, func(writer http.ResponseWriter, request *http.Request) {
-		writer.Header().Add("Content-Type", "application/xml")
-		// TODO: handle error
-		writer.Write(atomxml)
-	})
-	mux.HandleFunc(pathEntry, makePageHandler(pathEntry, viewHandler))
-	mux.Handle(pathStatic, )
-	mux.HandleFunc(pathIndex, makeIndexHandler())
+	// TODO: allow insecure traffic
+	mux, err := assets.NewCachedHandler(cfg.RootPath, true, cfg.Domain)
+	if err != nil {
+		log.WithFields(logrus.Fields{
+			logrus.ErrorKey: err,
+			"rootPath":      cfg.RootPath,
+			"domain":        cfg.Domain,
+		}).Fatal("can't start blog handler")
+	}
 
 	var globalHandler http.Handler
 	if len(cfg.Redirect) == 0 {